JETTY-592 reworked timers

+++ jetty/trunk/modules/server/jetty-util/src/main/java/org/mortbay/thread/Timeout.java
@@ -27,20 +27,27 @@ import org.mortbay.log.Log;
  * The nested class Task should be extended by users of this class to obtain call back notification of 
  * expiries. 
  * <p>
- * This class is not synchronized and the caller must protect against multiple thread access.
+ * This class is synchronized, but the callback to expired is not called within the synchronized scope.
  * 
  * @author gregw
  *
  */
 public class Timeout
 {
-    
+    private final Object _mutex;
     private long _duration;
     private long _now=System.currentTimeMillis();
     private Task _head=new Task();
-    
+
     public Timeout()
     {
+        _mutex=this;
+        _head._timeout=this;
+    }
+    
+    public Timeout(final Object mutex)
+    {
+        _mutex=mutex==null?this:mutex;
         _head._timeout=this;
     }
 
@@ -81,46 +88,35 @@ public class Timeout
     }
 
     /* ------------------------------------------------------------ */
-    public Task expired()
-    {
-        long _expiry = _now-_duration;
-        
-        if (_head._next!=_head)
-        {
-            Task task = _head._next;
-            if (task._timestamp>_expiry)
-                return null;
-            
-            task.unlink();
-            synchronized (task)
-            {
-                task._expired=true;
-            }
-            return task;
-        }
-        return null;
-    }
-
-    /* ------------------------------------------------------------ */
     public void tick()
     {
-        long _expiry = _now-_duration;
+        long expiry = _now-_duration;
         
-        while (_head._next!=_head)
+        while (true)
         {
-            Task task = _head._next;
-            if (task._timestamp>_expiry)
+            Task expired=null;
+            synchronized (_mutex)
+            {
+                if (_head._next!=_head && _head._next._timestamp<=expiry)
+                {
+                    expired=_head._next;
+                    expired.unlink();
+                    expired._expired=true;
+                }
+            }
+            if (expired!=null)
+            {
+                try
+                {
+                    expired.expired();
+                }
+                catch(Throwable th)
+                {
+                    Log.warn(Log.EXCEPTION,th);
+                }
+            }
+            else
                 break;
-            
-            task.unlink();
-            try
-            {
-                task.doExpire();
-            }
-            catch(Throwable th)
-            {
-                Log.warn(Log.EXCEPTION,th);
-            }
         }
     }
 
@@ -133,45 +129,63 @@ public class Timeout
     /* ------------------------------------------------------------ */
     public void schedule(Task task,long delay)
     {
-        if (task._timestamp!=0)
+        if (task._timeout!=null && task._timeout!=this)
         {
-            task.unlink();
-            task._timestamp=0;
+            task.cancel();
         }
-        task._expired=false;
-        task._delay=delay;
-        task._timestamp = _now+delay;
         
-        Task last=_head._prev;
-        while (last!=_head)
+        synchronized (_mutex)
         {
-            if (last._timestamp <= task._timestamp)
-                break;
-            last=last._prev;
+            if (task._timestamp!=0)
+            {
+                task.unlink();
+                task._timestamp=0;
+            }
+            task._expired=false;
+            task._delay=delay;
+            task._timestamp = _now+delay;
+
+            Task last=_head._prev;
+            while (last!=_head)
+            {
+                if (last._timestamp <= task._timestamp)
+                    break;
+                last=last._prev;
+            }
+            last.setNext(task);
         }
-        last.setNext(task);
     }
 
 
     /* ------------------------------------------------------------ */
     public void cancelAll()
     {
-        _head._next=_head._prev=_head;
+        synchronized (_mutex)
+        {
+            _head._next=_head._prev=_head;
+            // TODO call a cancel callback?
+        }
     }
 
     /* ------------------------------------------------------------ */
     public boolean isEmpty()
     {
-        return _head._next==_head;
+        synchronized (_mutex)
+        {
+            return _head._next==_head;
+        }
     }
 
     /* ------------------------------------------------------------ */
     public long getTimeToNext()
     {
-        if (_head._next==_head)
-            return -1;
-        long to_next = _duration+_head._next._timestamp-_now;
-        return to_next<0?0:to_next;
+        synchronized (_mutex)
+        {
+            if (_head._next==_head)
+                return -1;
+            long to_next = _duration+_head._next._timestamp-_now;
+            return to_next<0?0:to_next;
+        }
     }
 
     /* ------------------------------------------------------------ */
@@ -179,13 +193,16 @@ public class Timeout
     {
         StringBuilder buf = new StringBuilder();
         buf.append(super.toString());
-        
-        Task task = _head._next;
-        while (task!=_head)
+
+        synchronized (_mutex)
         {
-            buf.append("-->");
-            buf.append(task);
-            task=task._next;
+            Task task = _head._next;
+            while (task!=_head)
+            {
+                buf.append("-->");
+                buf.append(task);
+                task=task._next;
+            }
         }
         
         return buf.toString();
@@ -197,7 +214,7 @@ public class Timeout
     /* ------------------------------------------------------------ */
     /** Task.
      * The base class for scheduled timeouts.  This class should be
-     * extended to implement the expire() method, which is called if the
+     * extended to implement the {@link #expire()} or {@link #expired()} method, which is called if the
      * timeout expires.
      * 
      * @author gregw
@@ -211,29 +228,20 @@ public class Timeout
         long _delay;
         long _timestamp=0;
         boolean _expired=false;
-        Object _mutex=this;
 
+        /* ------------------------------------------------------------ */
         public Task()
         {
             _next=_prev=this;
         }
-        
-        public Task(Object mutex)
-        {
-            _next=_prev=this;
-            _mutex=mutex;
-        }
 
-        public void setMutex(Object mutex)
-        {
-            _mutex=mutex;
-        }
-
+        /* ------------------------------------------------------------ */
         public long getTimestamp()
         {
             return _timestamp;
         }
-        
+
+        /* ------------------------------------------------------------ */
         public long getAge()
         {
             Timeout t = _timeout;
@@ -241,8 +249,9 @@ public class Timeout
                 return t._now-_timestamp;
             return 0;
         }
-        
-        public void unlink()
+
+        /* ------------------------------------------------------------ */
+        private void unlink()
         {
             _next._prev=_prev;
             _prev._next=_next;
@@ -251,7 +260,8 @@ public class Timeout
             _expired=false;
         }
 
-        public void setNext(Task task)
+        /* ------------------------------------------------------------ */
+        private void setNext(Task task)
         {
             if (_timeout==null || 
                 task._timeout!=null && task._timeout!=_timeout ||    
@@ -266,47 +276,19 @@ public class Timeout
         }
         
         /* ------------------------------------------------------------ */
-        /** Schedule the task on the given timeout.
-         * The task exiry will be called after the timeout duration.
-         * @param timer
-         */
-        public void schedule(Timeout timer)
-        {
-            unlink();
-            timer.schedule(this);
-        }
-        
-        /* ------------------------------------------------------------ */
-        /** Schedule the task on the given timeout.
-         * The task exiry will be called after the timeout duration.
-         * @param timer
-         */
-        public void schedule(Timeout timer, long delay)
-        {
-            unlink();
-            timer.schedule(this,delay);
-        }
-        
-        /* ------------------------------------------------------------ */
-        /** Reschedule the task on the current timeout.
-         * The task timeout is rescheduled as if it had been canceled and
-         * scheduled on the current timeout.
-         */
-        public void reschedule()
-        {
-            Timeout timer = _timeout;
-            unlink();
-            timer.schedule(this,_delay);
-        }
-        
-        /* ------------------------------------------------------------ */
         /** Cancel the task.
          * Remove the task from the timeout.
          */
         public void cancel()
         {
-            _timestamp=0;
-            unlink();
+            if (_timeout!=null)
+            {
+                synchronized (_timeout._mutex)
+                {
+                    _timestamp=0;
+                    unlink();
+                }
+            }
         }
         
         /* ------------------------------------------------------------ */
@@ -317,19 +299,13 @@ public class Timeout
         
         /* ------------------------------------------------------------ */
         /** Expire task.
-         * This method is called when the timeout expires.
+         * This method is called when the timeout expires. It is called 
+         * outside of any synchronization scope and may be delayed. 
          * 
+         * @see #expire() For a synchronized callback.
          */
-        public void expire(){}
-        
-        private void doExpire()
-        {
-            synchronized (_mutex)
-            {
-                _expired=true;
-                expire();
-            }
-        }
+        public void expired(){}
+
     }
 
 }
